Revert "Remove base class BaseJupyterSession (#13824)"
This reverts commit 2628d701d530b60f5773f49ce4cf2ed9a6853ef0.
